feat: add Exception
