+++ src/main/java/com/sprint/mission/discodeit/exception/DatabaseExceptionAdvice.java
@@ -3,12 +3,15 @@ package com.sprint.mission.discodeit.exception;
 import java.time.Instant;
 import java.util.Map;
 import org.hibernate.exception.ConstraintViolationException;
+import org.springframework.core.Ordered;
+import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @RestControllerAdvice
+@Order(Ordered.HIGHEST_PRECEDENCE)
 public class DatabaseExceptionAdvice {
 
   @ResponseStatus(HttpStatus.CONFLICT)
@@ -49,17 +52,4 @@ public class DatabaseExceptionAdvice {
     }
   }
 
-  @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
-  @ExceptionHandler(Exception.class)
-  public ErrorResponse exception(Exception e) {
-
-    return ErrorResponse.builder()
-        .timestamp(Instant.now())
-        .code(ErrorCode.INTERNAL_SERVER_ERROR.getCode())
-        .message(e.getMessage())
-        .details(Map.of(e.getClass().getSimpleName(), "알 수 없는 서버 오류가 발생했습니다."))
-        .exceptionType(e.getClass().getSimpleName())
-        .status(HttpStatus.INTERNAL_SERVER_ERROR.value())
-        .build();
-  }
 }
