+++ src/main/java/com/sprint/mission/discodeit/exception/MessageExceptionAdvice.java
@@ -5,6 +5,8 @@ import com.sprint.mission.discodeit.exception.message.MessageNotFoundException;
 import com.sprint.mission.discodeit.exception.message.NullMessageContentException;
 import com.sprint.mission.discodeit.exception.message.NullMessageTitleException;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.core.Ordered;
+import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
@@ -12,6 +14,7 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @Slf4j
 @RestControllerAdvice
+@Order(Ordered.HIGHEST_PRECEDENCE)
 public class MessageExceptionAdvice {
 
   @ResponseStatus(HttpStatus.NOT_FOUND)
