+++ src/main/java/com/sprint/mission/discodeit/exception/ReadStatusExceptionAdvice.java
@@ -2,12 +2,15 @@ package com.sprint.mission.discodeit.exception;
 
 import com.sprint.mission.discodeit.exception.readstatus.ReadStatusExistsException;
 import com.sprint.mission.discodeit.exception.readstatus.ReadStatusNotFoundException;
+import org.springframework.core.Ordered;
+import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @RestControllerAdvice
+@Order(Ordered.HIGHEST_PRECEDENCE)
 public class ReadStatusExceptionAdvice {
 
   /**
